Simplify to single forecasting page
